+++ cmd/mq_prometheus/main.go
@@ -140,7 +140,7 @@ func main() {
 					discoverConfig.MetaPrefix = config.cf.MetaPrefix
 					err = mqmetric.DiscoverAndSubscribe(discoverConfig)
 					e := mqmetric.RediscoverAttributes(ibmmq.MQOT_CHANNEL, config.cf.MonitoredChannels)
-					e = mqmetric.RediscoverAttributes(mqmetric.OT_CHANNEL_MQTT, config.cf.MonitoredMQTTChannels)
+					_ = mqmetric.RediscoverAttributes(mqmetric.OT_CHANNEL_MQTT, config.cf.MonitoredMQTTChannels)
 
 					log.Debugf("Returned from RediscoverAttributes with error %v  tempErr %v", err, e)
 				}
@@ -205,6 +205,52 @@ func main() {
 	}
 }
 
+// Experimenting ...
+// See https://github.com/prometheus/exporter-toolkit/blob/master/web/handler.go
+// for an example of how we might add authentication via Basic Auth in here.
+//
+// This function lets us chain handlers, inserting our own processing on the way
+// before the prometheus handler does any work.
+// The basic_auth block in the prometheus engine yaml gives us a plaintext password that
+// can be crypted/hashed and compared to what's in a local config store. If it's OK, then
+// pass the request on to the real processor. Otherwise fail with a suitable HTTP code and
+// response. Since this is called on every scrape, we'd want to cache the results of the crypt.
+/*
+func myHandler(next http.HandlerFunc) http.HandlerFunc {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		authOK := true
+		u, p, auth := r.BasicAuth()
+		// log.Infof("   U: %s P: %s A:%v", u, p, auth)
+		//log.Infof("Handler headers: %v", r.Header)
+		//authHeader, ok := r.Header["Authorization"]
+		//if ok {
+		//	log.Infof("Auth header: %s", authHeader)
+		//} else {
+		//	log.Infof("Auth header: %s", "N/A")
+		//}
+		if auth {
+			authOK = validateUP(u, p)
+		}
+
+		if authOK {
+			next.ServeHTTP(w, r)
+		} else {
+			w.Header().Set("WWW-Authenticate", "Basic")
+			http.Error(w, http.StatusText(http.StatusUnauthorized), http.StatusUnauthorized)
+		}
+	})
+}
+// A rather simplistic validator!
+func validateUP(u string, p string) bool {
+	rc := true
+	if u != "user" || p != "password" {
+		rc = false
+	}
+
+	return rc
+}
+*/
+
 func startServer() {
 	var err error
 	// This function starts a new thread to handle the web server that will then run
@@ -212,9 +258,14 @@ func startServer() {
 
 	// Need to wait until signalled by the main thread that it's setup the gauges
 	log.Debug("HTTP server - waiting until MQ connection ready")
+
+	//myHandlerFunc := myHandler(promhttp.Handler().(http.HandlerFunc))
+
 	<-startChannel
 
 	http.Handle(config.httpMetricPath, promhttp.Handler())
+	//http.Handle(config.httpMetricPath, myHandlerFunc)
+
 	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
 		w.Write(landingPage())
 	})